CSV Reader - 4 byte delimiters (https://redirect.github.com/duckdb/duckdb/pull/14670)
This PR introduces support for delimiters of up to 4 bytes in the CSV
Reader. I chose 4 bytes primarily to allow 🦆 to be used as a delimiter
:-)

The parsing is implemented by adding intermediate states in the state
machine to handle these specific extra bytes.

The PR includes several tests to verify that repeated character patterns
(e.g., ABAC) are correctly parsed, and that multi-byte delimiters are
properly handled in cases where values span multiple buffers.

I'm happy to add more tests if anyone has additional ideas for
interesting edge cases to stress-test.

Upstream-commit: https://github.com/duckdb/duckdb/commit/ef43a0df34593f61dfaf81a40f92eb7bcbc220d8