[pystubs] Fix type of `proto` parameter in `from_substrait` methods. (https://redirect.github.com/duckdb/duckdb/pull/15004)
The generated signature for the two `from_substrait` methods make the
`proto` parameter a `str`. However, if called with a value of that type,
the following error is thrown:

```
TypeError: from_substrait(): incompatible function arguments. The following argument types are supported:
    1. (self: duckdb.duckdb.DuckDBPyConnection, proto: bytes) -> duckdb.duckdb.DuckDBPyRelation
```

This PR thus changes the type to `bytes` and adds a `stubgen` comment to
document why.

Upstream-commit: https://github.com/duckdb/duckdb/commit/165b5606ce5b9f041118e5be1036f1f9fd7c89de