Interrupt query on error in `ClientContext::Query` (https://redirect.github.com/duckdb/duckdb/pull/15058)
In the C++ API, when providing multiple statements to
`ClientContext::Query`, we would return a list of query results for each
of the statements. However, we would keep on executing all statements -
similar to e.g. the behavior you might expect if you copy-paste a list
of statements into a terminal.

In practice, many consumers of `ClientContext::Query` would ignore all
but the last result, and it is generally not desirable to continue
executing/swallow errors anyway. This PR switches the behavior to
instead stop when an error is encountered and return the result
containing the error.

Upstream-commit: https://github.com/duckdb/duckdb/commit/cd09b6e41e6efb5dd16e1317070420dfebe3ee0c