Pushdown inequality filters (https://redirect.github.com/duckdb/duckdb/pull/15059)
This PR extends the table filter pushdown to include inequality filters.
The support had been mostly there but they have just not been generated
by the optimizer. This enables row group pruning based on zonemaps and
early-pruning of scans of vectors using inequality filters. As with any
filter this is particularly effective when the filter is very selective.

Upstream-commit: https://github.com/duckdb/duckdb/commit/cbe809291b28394d494aa21472e694e9638cad39