Top-N: Perform global boundary checking before doing sort-key conversion (https://redirect.github.com/duckdb/duckdb/pull/15087)
This PR is essentially a partial revert of
https://github.com/duckdb/duckdb/pull/14424 - that fixes a performance
regression that was caused by the Top-N operator doing the global
boundary checking **after** constructing the sort key, instead of
before. This meant that, even for rows where we would be certain they
would not be added to the heap, we would be constructing sort keys. This
PR brings back the old code that allows the boundary values to be
checked directly on the sort keys. In order to facilitate this, we
deserialize the global boundary value (stored as a sort key) back into a
`DataChunk` that can be compared against directly.

Upstream-commit: https://github.com/duckdb/duckdb/commit/6064047fb295f723d55dd3f29aa4422852bd9383