Implement struct projection pushdown for JSON reads (https://redirect.github.com/duckdb/duckdb/pull/15116)
After @Mytherin did most of the heavy lifting in
https://github.com/duckdb/duckdb/pull/14750, I was able to implement
struct projection pushdown for JSON reads.

Although we still need to parse the entire JSON, we do not need to
convert it to DuckDB Vectors, which saves a lot of time. Here are the
results (not very scientific - some noise from my laptop doing other
stuff) when reading TPC-H SF1 directly from JSON files as a struct,
without struct projection pushdown (Before) and with (This PR).

As we can see, all queries are now faster. The difference is more
pronounced for some queries, e.g., Q3, than it is for others.

| Query | Before | This PR |
|:-|-:|-:|
| 1 | 0.665 | 0.478 |
| 2 | 0.234 | 0.183 |
| 3 | 0.813 | 0.209 |
| 4 | 0.774 | 0.601 |
| 5 | 0.890 | 0.582 |
| 6 | 0.617 | 0.425 |
| 7 | 0.807 | 0.433 |
| 8 | 0.898 | 0.588 |
| 9 | 0.970 | 0.616 |
| 10 | 0.808 | 0.551 |
| 11 | 0.179 | 0.150 |
| 12 | 0.785 | 0.603 |
| 13 | 0.258 | 0.208 |
| 14 | 0.702 | 0.478 |
| 15 | 0.655 | 0.449 |
| 16 | 0.200 | 0.170 |
| 17 | 1.372 | 0.848 |
| 18 | 1.457 | 0.903 |
| 19 | 0.703 | 0.482 |
| 20 | 0.767 | 0.525 |
| 21 | 2.143 | 1.495 |
| 22 | 0.319 | 0.240 |

Upstream-commit: https://github.com/duckdb/duckdb/commit/0b7b67dc84e82c9b4d61c1ad25287eb83829e55d