Parquet reader: rename metadata cache setting to `parquet_metadata_cache`, and avoid using it for stats (https://redirect.github.com/duckdb/duckdb/pull/15129)
This PR renames the parquet metadata cache setting to the more
descriptive `parquet_metadata_cache`:

```sql
SET parquet_metadata_cache=true;
```

In addition, we no longer use the metadata cache for propagating column
statistics. Previously, we would try to use metadata stored in the cache
to propagate column statistics - but this could cause large slowdowns
when scanning many files because (1) the code to obtain the stats from
the metadata was not optimized, and this could happen `number_of_files *
number_of_columns` times, and (2) we would open each file to check if
the file had not gone stale before doing this - which could again cause
many file system operations. This could cause large slowdowns when
enabling the metadata cache.

This PR just disables this behavior entirely and now uses the metadata
cache exclusively to prevent re-reading the metadata from disk. In a
future PR we might be able to use the stats again given that we can
greatly optimize the code for obtaining the stats.

Upstream-commit: https://github.com/duckdb/duckdb/commit/59b237ad1c9e59c7164be5b2728f2c2ca799bf5e