Add DictionaryId that can be used to uniquely identify dictionaries, and use this in the aggregate HT to cache look-ups (https://redirect.github.com/duckdb/duckdb/pull/15196)
Follow-up from https://github.com/duckdb/duckdb/pull/15152

This PR enables the caching of results computed for dictionaries across
vectors by (optionally) assigning them with a unique identifier. The
identifier is a string that needs to be unique so we can distinguish
between a vector having the same dictionary or a different dictionary as
the previous vector. It does not have to be globally unique (i.e. really
only in the same pipeline for the same thread).

* For our own storage, the dictionary id is set to the `ColumnSegment`
pointer address. Since we never de-allocate `ColumnSegments` while
scanning this works
* For Parquet dictionaries, the dictionary id is set to
`[parquet_filename]_[column_name]_[byte_offset]`

Upstream-commit: https://github.com/duckdb/duckdb/commit/2be6f64aedffdc52e81ee197828ac95126b1cc0f