NULL check for invalid result (https://redirect.github.com/duckdb/duckdb/pull/15194)
We run into a crash today where our code would check for column count in
a result and result in a crash. While we can keep track of this
ourselves and fix it, it may be helpful to have some checks in the C API
for DuckDB to check for the internal pointers being valid instead of a
crash.

Upstream-commit: https://github.com/duckdb/duckdb/commit/3b5feedfe7fd4354ecae907cf63c02c936e5e699