Rework `TableFilterType::CONSTANT_COMPARISON` to work identically to constant comparisons in SQL (https://redirect.github.com/duckdb/duckdb/pull/15197)
This PR reworks the `TableFilterType::CONSTANT_COMPARISON` so that it
works identically to constant comparisons in SQL with regards to NULL
values, i.e. it filters out all `NULL` values. Previously, this used to
be a bit of a mess:

* When encountering a constant comparison (e.g. `x = 5`), we would
always push `x=5 AND x IS NOT NULL`
* Constant comparisons in our table scan would remove `NULL` values, and
then run the `x IS NOT NULL` filter separately for no reason (since the
constant comparison had already removed all `NULL` values)
* Constant comparisons when converted to expressions would be converted
back into the `x = 5` form, which filters out `NULL` values, effectively
not being round-trippable
* For the Parquet reader, we would actually not filter out `NULL` values
in the constant comparison, but we would need to check if values were
`NULL` anyway - only to then always filter them out (since we never
pushed a constant comparison without a `IS NOT NULL` filter anyway)

This PR makes this consistent and speeds things up a little by skipping
unnecessary checks.

Upstream-commit: https://github.com/duckdb/duckdb/commit/1c6eacd28aae3f20337634b991d60ec001187313