Fix for row-id pushdown, and remove unnecessarily complicated method (https://redirect.github.com/duckdb/duckdb/pull/15216)
Row IDs don't have segments, so we can't do segment-level skipping.

This problem only shows up with dynamic filters, since the filter
changes while we're in the middle of a scan. With a static filter we
would never get to this point (since the row group would be filtered out
entirely at an earlier stage).

Upstream-commit: https://github.com/duckdb/duckdb/commit/087d8f34f6d974a1fedfc500fb5c8e269da01c95