Rework upload ci via reusable workflow (https://redirect.github.com/duckdb/duckdb/pull/15243)
This changes R.yml, Wasm.yml and OSX.yml to upload the duckdb extensions
built there to go through the reusable workflow already introduced for
Linux extensions.

Advantage in centralizing to single workflow is that it makes easier to
track and modify it.

Also the change make from steps executed only on nightly CI to step
always executed, and only the actual upload will fail on PRs. This
should make so this PR CI can be reviewed and have some guarantees that
stuff should happening as intended.

This has a risk of breaking upload on nightly (for OSX and R), but I
think this is made in way to minimize potential problems.
For Wasm this make so that nigthly duckdb version get the equivalent set
of `duckdb_extension.wasm`.

Upstream-commit: https://github.com/duckdb/duckdb/commit/82abc6a40de7ed102c86d12ed5ec09bfa207d559