Implement https://redirect.github.com/duckdb/duckdb/pull/14513: implement support for (a, b) IN (SELECT ...) for uncorrelated subqueries (https://redirect.github.com/duckdb/duckdb/pull/15259)
Implement https://redirect.github.com/duckdb/duckdb/pull/14513 

The following syntax is now supported **for uncorrelated subqueries**,
following what also works in Postgres:

```sql
SELECT (1, 2) IN (SELECT a, b FROM tbl);
```

The planning also works correctly for correlated subqueries - but the
mark join with multiple join conditions is not correctly executed in the
hash join currently (see the FIXME in the code). Should be addressed in
the future but I felt this change was worthwhile by itself.

Upstream-commit: https://github.com/duckdb/duckdb/commit/c828cadfc1c5f8df89f94cdb8eb62aba8fdb5e5e