remove O_SYNC from O_DIRECT (https://redirect.github.com/duckdb/duckdb/pull/15294)
The `config.use_direct_io` option has a surprising effect, which is that
it also enables `O_SYNC`. This is probably not the expected or intended
behavior of use_direct_io. This PR removes O_SYNC. It also re-enables
direct_io for osx. I don't have context on why that was disabled so
perhaps there was a good reason to disable it.

As I understand things:
- O_DIRECT disables the page-cache. This makes good sense for DuckDB,
since DuckDB already has a page cache.
- O_SYNC forces all writes to be durable before they complete. This can
significantly reduce performance on slower media (HDD, NFS, etc.). This
doesn't appear to add any value for DuckDB since DuckDB uses fsync() for
write ordering regardless.

Upstream-commit: https://github.com/duckdb/duckdb/commit/4835faaa9288a75ebbc36811f7142fbb2dd7ed87