Use correct element rename_list_el in grammar (https://redirect.github.com/duckdb/duckdb/pull/15339)
This was the actual definition used by the generated grammar - but
somehow the generated grammar ended up differing from the checked in
grammar. We should add the grammar to the generated files check. The
only problem there is installing the correct version of flex/bison in
the CI:

```
$ > bison --version
bison (GNU Bison) 2.3
Written by Robert Corbett and Richard Stallman.

Copyright (C) 2006 Free Software Foundation, Inc.
This is free software; see the source for copying conditions.  There is NO
warranty; not even for MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.
$ > flex --version
flex 2.6.4 Apple(flex-34)
```

Upstream-commit: https://github.com/duckdb/duckdb/commit/b9bea84a9ece056876634681b673740bd6a283b2