Append default to appender (https://redirect.github.com/duckdb/duckdb/pull/15121)
The PRs [Support default values when appending data
chunks](https://github.com/duckdb/duckdb/pull/14733) and [C API -
Implement AddColumn and ClearColumns for the
Appender](https://github.com/duckdb/duckdb/pull/14880/) added support
for appending to columns with default values when using
`duckdb_append_data_chunk` but the API is very verbose compared to
[duckdb_append_default](https://duckdb.org/docs/api/c/appender#duckdb_append_default)
function that works with the non-chunk approach for appending values.

For example, if a table has 10 columns and only 2 columns have default
values, we will need to call `duckdb_appender_add_column` 8 times with
the names of those eight columns. Another problem is that if we
sometimes want to insert default values for those 2 columns and
sometimes insert some specific values, we will have to call
`duckdb_appender_clear_columns` and recreate data chunk to insert
specific values in columns that have default values.

I propose adding a new function, `duckdb_append_default_to_chunk` to
solve these pain points and have a simpler API that solves both of these
issues. With this function, you can assign a default value to the data
chunk for the specified column, and you do not need to recreate the
chunk if you want to use default value for some rows and non-default for
others.

I implemented it in this PR and it worked fine for me in the limited
tests that I ran. If you think this is useful I can add the missing
tests and add it to the header generation files.

Thoughts? @taniabogatsch @Mytherin 

Upstream-commit: https://github.com/duckdb/duckdb/commit/6ce195704a8d3af894766af51e0451955e4fc93d